requirements.txt
fastapi
uvicorn
openai
anthropic
google-genai
python-dotenv
